Config now loaded as a module, prevents namespace clashing
Original patch by Chris Brannon.  I modified it a bit to clean out some
leftover things that weren't needed because of this patch.

This has the nasty side effect of compiling .slurpyrc into .slurpyrcc,
but the added safety of not exec'ing every line in the config and the
flexibility of still using python as the configuration outweigh this.